Refactor ExtBuilders
